# platform/swaggo/swagger.yaml
basePath: /flighthours/api/v1
definitions:
...
  handlers.CreateDailyLogbookRequest:
    properties:
      book_page:
        type: integer
      log_date:
        type: string
    required:
    - log_date
    type: object
...
  handlers.DailyLogbookDeleteResponse:
    properties:
      _links:
        items:
          $ref: '#/definitions/handlers.Link'
        type: array
      deleted:
        type: boolean
      id:
        type: string
    type: object
...
  handlers.DailyLogbookListResponse:
    properties:
      _links:
        items:
          $ref: '#/definitions/handlers.Link'
        type: array
      daily_logbooks:
        items:
          $ref: '#/definitions/handlers.DailyLogbookResponse'
        type: array
      total:
        type: integer
    type: object
  handlers.DailyLogbookResponse:
    properties:
      _links:
        items:
          $ref: '#/definitions/handlers.Link'
        type: array
      book_page:
        type: integer
      employee_id:
        type: string
      id:
        type: string
      log_date:
        type: string
      status:
        type: string
    type: object
  handlers.DailyLogbookStatusResponse:
    properties:
      _links:
        items:
          $ref: '#/definitions/handlers.Link'
        type: array
      id:
        type: string
      status:
        type: string
      updated:
        type: boolean
    type: object
...
  handlers.Link:
    properties:
      href:
        type: string
      method:
        type: string
      rel:
        type: string
    type: object
...
  handlers.UpdateDailyLogbookRequest:
    properties:
      book_page:
        type: integer
      log_date:
        type: string
      status:
        type: boolean
    required:
    - log_date
    type: object
...
  middleware.APIResponse:
    properties:
      code:
        type: string
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
host: localhost:8081
info:
  contact:
    email: support@champion19.com
    name: Champion19 Support
  description: API para gestión de horas de vuelo de pilotos y tripulantes
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Flighthours API
  version: "1.0"
paths:
...
  /daily-logbooks:
    get:
      description: Returns a list of daily logbooks for the currently authenticated
        employee
      parameters:
      - description: Filter by status (true for active, false for inactive)
        in: query
        name: status
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.DailyLogbookListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middleware.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.APIResponse'
      security:
      - BearerAuth: []
      summary: List daily logbooks for authenticated employee
      tags:
      - DailyLogbooks
    post:
      consumes:
      - application/json
      description: Creates a new daily logbook for the authenticated employee
      parameters:
      - description: Daily logbook data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateDailyLogbookRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.DailyLogbookResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middleware.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middleware.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.APIResponse'
      security:
      - BearerAuth: []
      summary: Create a new daily logbook
      tags:
      - DailyLogbooks
  /daily-logbooks/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes an existing daily logbook for the authenticated employee
        (accepts both UUID and obfuscated ID)
      parameters:
      - description: Daily Logbook ID (obfuscated ID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.DailyLogbookDeleteResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middleware.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middleware.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/middleware.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/middleware.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete a daily logbook
      tags:
      - DailyLogbooks
    get:
      consumes:
      - application/json
      description: Returns daily logbook information by ID for the authenticated employee
        (accepts both UUID and obfuscated ID)
      parameters:
      - description: Daily Logbook ID (obfuscated ID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.DailyLogbookResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middleware.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middleware.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/middleware.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/middleware.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.APIResponse'
      security:
      - BearerAuth: []
      summary: Get daily logbook by ID
      tags:
      - DailyLogbooks
    put:
      consumes:
      - application/json
      description: Updates an existing daily logbook for the authenticated employee
        (accepts both UUID and obfuscated ID)
      parameters:
      - description: Daily Logbook ID (obfuscated ID)
        in: path
        name: id
        required: true
        type: string
      - description: Updated daily logbook data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateDailyLogbookRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.DailyLogbookResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middleware.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middleware.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/middleware.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/middleware.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.APIResponse'
      security:
      - BearerAuth: []
      summary: Update a daily logbook
      tags:
      - DailyLogbooks
  /daily-logbooks/{id}/activate:
    patch:
      consumes:
      - application/json
      description: Sets daily logbook status to active (accepts both UUID and obfuscated
        ID)
      parameters:
      - description: Daily Logbook ID (obfuscated ID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.DailyLogbookStatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middleware.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middleware.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/middleware.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/middleware.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.APIResponse'
      security:
      - BearerAuth: []
      summary: Activate a daily logbook
      tags:
      - DailyLogbooks
  /daily-logbooks/{id}/deactivate:
    patch:
      consumes:
      - application/json
      description: Sets daily logbook status to inactive (accepts both UUID and obfuscated
        ID)
      parameters:
      - description: Daily Logbook ID (obfuscated ID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.DailyLogbookStatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middleware.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middleware.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/middleware.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/middleware.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.APIResponse'
      security:
      - BearerAuth: []
      summary: Deactivate a daily logbook
      tags:
      - DailyLogbooks
...
securityDefinitions:
  BearerAuth:
    description: 'JWT Bearer token (format: "Bearer {token}")'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"